+++ flake.nix
@@ -21,7 +21,7 @@
       in {
         packages.default = pkgs.rustPlatform.buildRustPackage {
           pname = "hoogle-tui";
-          version = "0.1.0";
+          version = "0.1.1";
           src = ./.;
           cargoLock.lockFile = ./Cargo.lock;